restructed forgot pin files
