+++ src/app/auth/forgot-pin/reset-pin/page.tsx
