+++ src/app/auth/forgot-pin/reset-pin/sections/submit.tsx
@@ -2,26 +2,34 @@
 
 import { useAtomValue } from "jotai";
 import Button from "@/components/button";
+import { usePhoneAuth } from "@/api/auth";
 import { useRouter } from "next/navigation";
-import { useOnboardOTPVerify } from "@/api/auth";
 import { authAtom } from "@/lib/atoms/auth.atom";
+import { TPhoneAuthResponse } from "@/api/auth/types";
+import phoneValidator from "validator/lib/isMobilePhone";
 
 const Submit = () => {
   const router = useRouter();
   const form = useAtomValue(authAtom);
-  const { mutate: verify, isPending, isSuccess } = useOnboardOTPVerify();
+  const { isPending, isSuccess, mutate: request } = usePhoneAuth();
 
   const isLoading = isPending || isSuccess;
-  const isFormInvalid = form.code.length < 6 || !form.phoneNumber;
-  const isSubmitDisabled = isFormInvalid || isLoading;
 
-  const onSuccess = () => {
-    router.push("/auth/create-new-pin");
+  const isValidPhoneNumber =
+    !!form.phoneNumber.trim() && phoneValidator(form.phoneNumber);
+
+  const isFormInvalid = !isValidPhoneNumber;
+  const isSubmitDisabled = isLoading || isFormInvalid;
+
+  const onSuccess = (data: TPhoneAuthResponse) => {
+    if (data?.exists) {
+      router.push("/auth/forgot-pin/verify");
+    }
   };
 
   const handleSubmit = () => {
     if (isSubmitDisabled) return;
-    verify({ code: form.code, phone: form.phoneNumber }, { onSuccess });
+    request({ phone: form.phoneNumber }, { onSuccess });
   };
 
   return (
