+++ src/app/auth/forgot-pin/verify/sections/components.tsx
